pull updates for tag_utility into master
Enhancements were made to tag utility in the course of working on
the vmc_sonos_customizations stuff: there was a bit of refactoring
and the interactive edit feature was completed.

The test_sound.mp3 file was added just to avoid having to manually add
it when testing/developing. (laziness, sorry)
	modified:   tag_utility
	new file:   test_sound.mp3
	new file:   ../lib/media_editor.rb
	modified:   ../lib/tag_munger.rb